backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "google>=3.0.0",
    "google-genai>=1.66.0",
    "numpy>=2.4.2",
    "openai>=2.26.0",
    "pydantic-settings>=2.13.1",
    "pydub>=0.25.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "soundfile>=0.13.1",
    "supabase>=2.28.0",
]
